This   system checks and reports changes in router configs via call/mail/text
This project checks the existing configuration against a pre-defined
configuration. This role can also be on the basis of service that the
router is supporting as in if a router is part (role) of RIP scheme, its
startup configuration will be different from router which is running
OSPF.
The role based nature of the project allows the user to only run the
compliance test but only the network admin can run the test in addition
to viewing the existing users, adding/deleting users and exiting the
application.

Objectives:
1) Role Based privileges i.e report visibility based on organizational
structure.
2) Check the existing configuration against a pre-defined configuration.
This checking can be a role based checking as in
if a router is part (role) of RIP scheme, its startup configuration will
be different from router which is running OSPF.
This role can also be on the basis of service that the router is
supporting.
3) SLA can be attributed for routers if it does not comply with the
expected startup configuration.
4) Graphical representation showing the non-compliant routers in red.
5) Automated email will be sent at end of a compliance check cycle
indicating compliance and non-compliance statistics.
6) Report logging with time stamping as per cycle mentioned by user

Building steps of the project:
1. Foremost, the graphic user interface for the router compliance test
is called using welcome.py file and the TKinter module.

2. login.py file creates the interface to enquire the login credentials
of both the network admin and the users.

3. database_creator.py validates the login credentials and displays the
options corresponding to the user and the network admin using the eg.py
file. It returns the message “incorrect credentials” on entering data
not in the login database.

4. The initiate.py file runs the router compliance test and returns the
action test list. This list has the flags corresponding to each of the
routers; 0 corresponds to an non-compliant router and 1 corresponds to a
compliant router. The 1st value, 2nd value and 3rd value correspond to
the 1st, 3rd and 2nd router respectively.
4. The initiate.py file runs the router compliance test and returns the
action test list. This list has the flags corresponding to each of the
routers; 0 corresponds to an non-compliant router and 1 corresponds to a
compliant router. The 1st value, 2nd value and 3rd value correspond to
the 1st, 3rd and 2nd router respectively.
If the action set is [1,1,1], the SLA violation is displayed using the
green circle.

5. The networkx graph is also created using the initiate.py file. Each
router is labelled with its management IP. The green routers are the
compliant routers and the red routers are the non-compliant routers.

6. The twilio.py module sends text message to the network admin
describing SLA compliance and SLA violation. In case of violation, the
message identifies the non-compliant router and also mentions the
details of the configuration mismatch.

7. The initiate.py module sends e-mail to the network admin describing
SLA compliance and SLA violation. In case of violation, the message
identifies the non-compliant router and also mentions the details of the
configuration mismatch. It also sends the networkx graph as an
attachment